Move the historical cases chart
