+++ src/components/HistoricalCasesChart/BarChartTooltip.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 
-import ChartTooltip from "../../components/design-system/ChartTooltip";
-import Colors from "../../components/design-system/Colors";
-import { DateMMMMdyyyy } from "../../components/design-system/DateFormats";
-import { ModelInputs } from "../types";
+import { ModelInputs } from "../../page-multi-facility/types";
+import ChartTooltip from "../design-system/ChartTooltip";
+import Colors from "../design-system/Colors";
+import { DateMMMMdyyyy } from "../design-system/DateFormats";
 
 type MissingDataInput = {
   missing: boolean;
